+++ lib/secret_config.rb
@@ -7,9 +7,11 @@ require "secret_config/railtie" if defined?(Rails)
 # Centralized Configuration and Secrets Management for Ruby and Rails applications.
 module SecretConfig
   # When a node is both a value and a hash/branch in the tree, put its value in its hash with the following key:
-  NODE_KEY = "__value__".freeze
-  FILTERED = "[FILTERED]".freeze
-  RANDOM   = "$(random)".freeze
+  NODE_KEY   = "__value__".freeze
+  # A node with the following key imports the settings under the path it refers to into its parent node.
+  IMPORT_KEY = "__import__".freeze
+  FILTERED   = "[FILTERED]".freeze
+  RANDOM     = "$(random)".freeze
 
   module Providers
     autoload :File, "secret_config/providers/file"